pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ha-tools"
version = "0.1.0"
description = "High-performance CLI tool for AI agents working with Home Assistant configurations"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Home Assistant Tools", email = "tools@home-assistant.io"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.11"
dependencies = [
    "typer[rich]>=0.12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "aiohttp>=3.10.0",
    "asyncmy>=0.2.9",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.19.0",
    "PyYAML>=6.0.1",
    "rich>=13.7.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "types-PyYAML",
    "pre-commit>=3.6.0",
]

[project.scripts]
ha-tools = "ha_tools.cli:cli_main"

[project.urls]
Homepage = "https://github.com/home-assistant/ha-tools"
Documentation = "https://github.com/home-assistant/ha-tools#readme"
Repository = "https://github.com/home-assistant/ha-tools.git"
Issues = "https://github.com/home-assistant/ha-tools/issues"

[tool.hatch.build.targets.wheel]
packages = ["ha_tools"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["asyncmy.*", "asyncpg.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
asyncio_default_test_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=ha_tools --cov-report=term-missing"
